Setting EOS only for the last bulk
Previously EOS flag in bulk transform was set right after the last message from
kafka was received. However, bulks of messages received from kafka do not
correspond to GenStage bulks emitted, therefore, GenStage bulks could have been
mistakenly set as last.

This could occurr in the following situation. Let there be 1000 messages in
kafka topic. Consumer fetches them all in single bulk and puts them to queue.
Since the Consumer is at the end of the topic it sets the `is_end_of_stream` to
`true`. The GenStage demand is 500, thus 500 messages is dequeued and bulk
transformer is called with `is_end_of_stream = true`. However, this is
incorrect, since this bulk is not at the end of stream, there are still messages
waiting in the queue.

The fix adds another mechanism to determine whether GenStage bulk is to be marked as end of stream.
The original flag is still used, but only for empty stream detection (i.e. transforming empty bulk,
because there are no messages to be processed). For non-empty bulks a new mechanism was added. There
is now a queue tracking offsets where end of stream occurred when receiving kafka message
sets. When receiving kafka message set which is last the offset of the last message is stored in the
queue. When GenStage bulks are created later the queue is used as reference and end of stream is
set for bulks containing the offsets from the queue.